// admin/src/utils/types.ts
export interface User {
    id: string
    firstName: string
    lastName: string
    email: string
    admin: boolean
    phone: string | undefined
    createdAt: string
    updatedAt: string
}

export interface Team {
    id: string
    name: string
    description: string
    managerId: string
    startHour: number
    endHour: number
    createdAt: string
    updatedAt: string
}

export interface Clock {
    id: string
    user_id: string
    at: string
}

export interface LeavePeriod {
    id: string
    user_id: string
    startDate: string
    endDate: string
    accepted: boolean
    createdAt: string
}